requirements.txt
click==5.1
compass-stylesheets==0.12.6
Django==1.10.4
django-mustache==1.1.0
djangorestframework==3.5.3
https://github.com/sheppard/django-media-thumbnailer/archive/master.zip
FormEncode==2.0.0a1
html-json-forms==1.0.0
linecache2==1.0.0
Markdown==2.6.7
natural-keys==1.1.0
olefile==0.43
Pillow==4.0.0
pkg-resources==0.0.0
psycopg2==2.6.2
pyScss==1.3.5
pystache==0.5.4
pyxform==0.9.24
PyYAML==3.12
requests==2.12.4
six==1.10.0
swapper==1.0.0
traceback2==1.4.0
unicodecsv==0.14.1
unittest2==1.1.0
https://github.com/wq/xlsconv/archive/master.zip
https://github.com/wq/wq.app/archive/master.zip
wq.core==1.0.0b1
https://github.com/wq/wq.db/archive/master.zip
wq.io==1.0.0b1
https://github.com/wq/wq.start/archive/master.zip
xlrd==1.0.0
